Split out VirtTest utilities
Some of the methods in legacy `VirtTest` can be used in nrunner as well.
Because the legacy runner and nrunner will be both available for some
time, let's split it out of the same class into a common utility.
